base de datos a la perfeccion
y otras cosillas que dificultaban la apertura de pantallas, ahora estan todas bien, y podemos ver la puntuacion online en tiempo real. Falta refractorizar obviament

+++ core/src/com/mygdx/ttrispo/Pantalla/Partida.java
@@ -19,7 +19,7 @@ public class Partida extends PantallaBase {
     private GestorPiezas gestorPiezas;
     public static float posicionX, posicionY;
     private Procesador procesador;
-    private static int puntuacion;
+    private static long puntuacion;
 
     public Partida(MyGdxGame game) {
         super(game);
@@ -183,7 +183,7 @@ public class Partida extends PantallaBase {
         return gestorPiezas.getTexturaBloque(tipo);
     }
     
-    public int getPuntuacion(){
+    public long getPuntuacion(){
         return this.puntuacion;
     }
